pyproject.toml
[tool.poetry]
name = "test-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.0"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
drf-yasg = "^1.21.7"
setuptools = "^70.0.0"
django-filter = "^24.2"
django-cors-headers = "^4.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
